+++ returns/pointfree.py
@@ -1,3 +1,4 @@
+from returns._generated.pointfree.apply import _apply as apply
 from returns._generated.pointfree.bind import _bind as bind
 from returns._generated.pointfree.rescue import _rescue as rescue
 from returns._generated.pointfree.unify import _unify as unify